feat(cloud): expose connectProvider() in @agent-relay/cloud SDK
Lifts the Daytona-brokered provider auth flow out of the relay CLI and into
@agent-relay/cloud so other CLIs (ricky, future tools) can drive it directly
via `import { connectProvider } from '@agent-relay/cloud'` instead of shelling
out to `agent-relay cloud connect`.

- New exports: connectProvider, getProviderHelpText, normalizeProvider,
  runInteractiveSession, formatShellInvocation, wrapWithLaunchCheckpoint,
  loadSSH2, createAskpassScript, buildSystemSshArgs, plus SSH-related types.
- ssh2 is now an optionalDependency of @agent-relay/cloud — the package falls
  back to system ssh when ssh2 isn't installed (existing behavior, just newly
  exposed to consumers).
- Cloud SDK bumped 6.0.2 → 6.1.0.
- The `agent-relay cloud connect` command body shrinks from ~150 LOC to a
  thin wrapper around connectProvider() that keeps telemetry attribution.
- src/cli/lib/{ssh-interactive,auth-ssh}.ts now re-export the SDK helpers
  (rather than redefining them) so the existing `auth` command path is
  unchanged. ssh-interactive.test.ts moved to packages/cloud alongside
  the implementation.
