+++ package.json
@@ -130,7 +130,7 @@
   },
   "homepage": "https://github.com/AgentWorkforce/relay#readme",
   "dependencies": {
-    "@agent-relay/cloud": "6.0.2",
+    "@agent-relay/cloud": "6.1.0",
     "@agent-relay/config": "6.0.2",
     "@agent-relay/hooks": "6.0.2",
     "@agent-relay/sdk": "6.0.2",
