+++ packages/cloud/src/index.ts
@@ -5,14 +5,14 @@ export {
   refreshStoredAuth,
   ensureAuthenticated,
   authorizedApiFetch,
-} from "./auth.js";
+} from './auth.js';
 
 export {
   CloudApiClient,
   buildApiUrl,
   type CloudApiClientOptions,
   type CloudApiClientSnapshot,
-} from "./api-client.js";
+} from './api-client.js';
 
 export {
   runWorkflow,
@@ -23,7 +23,33 @@ export {
   resolveWorkflowInput,
   inferWorkflowFileType,
   shouldSyncCodeByDefault,
-} from "./workflows.js";
+} from './workflows.js';
+
+export {
+  connectProvider,
+  getProviderHelpText,
+  normalizeProvider,
+  type ConnectProviderIo,
+  type ConnectProviderOptions,
+  type ConnectProviderResult,
+} from './connect.js';
+
+export {
+  runInteractiveSession,
+  formatShellInvocation,
+  wrapWithLaunchCheckpoint,
+  type SshConnectionInfo,
+  type InteractiveSessionOptions,
+  type InteractiveSessionResult,
+} from './lib/ssh-interactive.js';
+
+export {
+  loadSSH2,
+  createAskpassScript,
+  buildSystemSshArgs,
+  DEFAULT_SSH_RUNTIME,
+  type AuthSshRuntime,
+} from './lib/ssh-runtime.js';
 
 export {
   type StoredAuth,
@@ -38,4 +64,4 @@ export {
   AUTH_FILE_PATH,
   defaultApiUrl,
   isSupportedProvider,
-} from "./types.js";
+} from './types.js';
